+++ include/H2DE/H2DE.h
@@ -7,7 +7,7 @@
  * Hydren 2 Dimensional Engine
  * https://github.com/Hydrened/H2DE
  * 
- * Version: 3.2.3
+ * Version: 3.2.4
  * 
  * Copyright (C) 2024-2025
  */